+++ srunner/tools/openscenario_parser.py
@@ -476,6 +476,10 @@ class OpenScenarioParser(object):
                         atomic = atomic_cls(trigger_actor, other_actor_type=triggered_type,
                                             terminate_on_failure=True, name=condition_name)
 
+                    else:
+                        atomic_cls = py_trees.meta.inverter(CollisionTest)
+                        atomic = atomic_cls(trigger_actor, terminate_on_failure=True, name=condition_name)
+
                 elif entity_condition.find('OffroadCondition') is not None:
                     off_condition = entity_condition.find('OffroadCondition')
                     condition_duration = float(off_condition.attrib.get('duration'))
@@ -483,7 +487,32 @@ class OpenScenarioParser(object):
                     atomic = atomic_cls(
                         trigger_actor, condition_duration, terminate_on_failure=True, name=condition_name)
                 elif entity_condition.find('TimeHeadwayCondition') is not None:
-                    raise NotImplementedError("TimeHeadway conditions are not yet supported")
+                    headtime_condition = entity_condition.find('TimeHeadwayCondition')
+
+                    condition_value = float(headtime_condition.attrib.get('value'))
+
+                    condition_rule = headtime_condition.attrib.get('rule')
+                    condition_operator = OpenScenarioParser.operators[condition_rule]
+
+                    condition_freespace = strtobool(headtime_condition.attrib.get('freespace', False))
+                    if condition_freespace:
+                        raise NotImplementedError(
+                            "TimeHeadwayCondition: freespace attribute is currently not implemented")
+                    condition_along_route = strtobool(headtime_condition.attrib.get('alongRoute', False))
+
+                    for actor in actor_list:
+                        if headtime_condition.attrib.get('entityRef', None) == actor.attributes['role_name']:
+                            triggered_actor = actor
+                            break
+                    if triggered_actor is None:
+                        raise AttributeError("Cannot find actor '{}' for condition".format(
+                            headtime_condition.attrib.get('entityRef', None)))
+
+                    atomic = InTimeToArrivalToVehicle(
+                        trigger_actor, triggered_actor, condition_value,
+                        condition_along_route, condition_operator, condition_name
+                    )
+
                 elif entity_condition.find('TimeToCollisionCondition') is not None:
                     ttc_condition = entity_condition.find('TimeToCollisionCondition')
 
@@ -493,10 +522,16 @@ class OpenScenarioParser(object):
                     condition_value = ttc_condition.attrib.get('value')
                     condition_target = ttc_condition.find('TimeToCollisionConditionTarget')
 
+                    condition_freespace = strtobool(ttc_condition.attrib.get('freespace', False))
+                    if condition_freespace:
+                        raise NotImplementedError(
+                            "TimeToCollisionCondition: freespace attribute is currently not implemented")
+                    condition_along_route = strtobool(ttc_condition.attrib.get('alongRoute', False))
+
                     if condition_target.find('Position') is not None:
                         position = condition_target.find('Position')
                         atomic = InTimeToArrivalToOSCPosition(
-                            trigger_actor, position, condition_value, condition_operator)
+                            trigger_actor, position, condition_value, condition_along_route, condition_operator)
                     else:
                         for actor in actor_list:
                             if ttc_condition.attrib.get('EntityRef', None) == actor.attributes['role_name']:
@@ -507,7 +542,8 @@ class OpenScenarioParser(object):
                                 ttc_condition.attrib.get('EntityRef', None)))
 
                         atomic = InTimeToArrivalToVehicle(
-                            trigger_actor, triggered_actor, condition_value, condition_operator)
+                            trigger_actor, triggered_actor, condition_value,
+                            condition_along_route, condition_operator, condition_name)
                 elif entity_condition.find('AccelerationCondition') is not None:
                     accel_condition = entity_condition.find('AccelerationCondition')
                     condition_value = float(accel_condition.attrib.get('value'))
@@ -556,17 +592,32 @@ class OpenScenarioParser(object):
                         trigger_actor, position, distance_value, name=condition_name)
                 elif entity_condition.find('DistanceCondition') is not None:
                     distance_condition = entity_condition.find('DistanceCondition')
+
                     distance_value = float(distance_condition.attrib.get('value'))
+
                     distance_rule = distance_condition.attrib.get('rule')
                     distance_operator = OpenScenarioParser.operators[distance_rule]
+
+                    distance_freespace = strtobool(distance_condition.attrib.get('freespace', False))
+                    if distance_freespace:
+                        raise NotImplementedError(
+                            "DistanceCondition: freespace attribute is currently not implemented")
+                    distance_along_route = strtobool(distance_condition.attrib.get('alongRoute', False))
+
                     if distance_condition.find('Position') is not None:
                         position = distance_condition.find('Position')
                         atomic = InTriggerDistanceToOSCPosition(
-                            trigger_actor, position, distance_value, distance_operator, name=condition_name)
+                            trigger_actor, position, distance_value, distance_along_route,
+                            distance_operator, name=condition_name)
 
                 elif entity_condition.find('RelativeDistanceCondition') is not None:
                     distance_condition = entity_condition.find('RelativeDistanceCondition')
                     distance_value = float(distance_condition.attrib.get('value'))
+
+                    distance_freespace = strtobool(distance_condition.attrib.get('freespace', False))
+                    if distance_freespace:
+                        raise NotImplementedError(
+                            "RelativeDistanceCondition: freespace attribute is currently not implemented")
                     if distance_condition.attrib.get('relativeDistanceType') == "cartesianDistance":
                         for actor in actor_list:
                             if distance_condition.attrib.get('entityRef', None) == actor.attributes['role_name']: