+++ Docs/CHANGELOG.md
@@ -23,7 +23,11 @@
 * The new weather parameters (related to fog) are now correctly read when running scenarios outside routes.
 * Enable weather animation during scenario execution (requires ephem pip package)
 * OpenSCENARIO support:
-    - Added support for controllers and provided default implementations for vehicles and pedestrians. This required changing the handling of actors, which results in that now all actors are controlled by an OSC controller.
+    - Added support for controllers and provided default implementations for vehicles and pedestrians. This required changing the handling of actors, which results in that now all actors are controlled by an OSC controller. Supported controllers:
+        - Pedestrian controller
+        - NPC vehicle controller (based on CARLA LocalPlanner)
+        - Simple vehicle controller to set velocities not brake/throttle
+        - External controller (to forward control to external entities)
     - Added initial speed support for pedestrians for OpenSCENARIO
     - Support for EnvironmentActions within Story (before only within Init). This allows changing weather conditions during scenario execution
     - Added support for RelativeSpeedCondition
@@ -32,12 +36,15 @@
     - Added support for OffroadCondition
     - Added support for CollisionCondition
     - Added support for EndOfRoadCondition
+    - Added support for TimeHeadwayCondition
     - Extended FollowLeadingVehicle example to illustrate weather changes
     - Created example scenarios to illustrate usage of controllers and weather changes
     - Reworked the handling of Catalogs to make it compliant to the 1.0 version (relative paths have to be relative to the scenario file)
     - The RoadNetwork can be defined as global Parameter
     - Fixed handling of relative positions with negative offset
     - Added support for local ParamaterDeclarations
+    - Fixed use of relative initial positions for any actor
+    - Added possibility to use synchronous execution mode with OpenSCENARIO
 * Atomics:
     - Several new atomics to enable usage of OSC controllers
     - WeatherBehavior to simulate weather over time
@@ -50,6 +57,8 @@
     - Added new *EndofRoadTest* criteria, to detect when a vehicle changes between OpenDRIVE roads.
     - CollisionTest criterion can now filter the collisions for a specific actor, or actor type_id.
     - Added a *duration* argument to *OnSidewalkTest* criteria, which makes the criteria fail after a certain time has passed, instead of doing so immediately. The default behavior has been unchanged.
+    - InTimeToArrivalToVehicle has had its two actor arguments swapped, to match all the other behaviors.
+    - Added *along_route* flag to InTimeToArrivalToVehicle, to take into account the topology of the road
 * Removed unsupported scenarios (ChallengeBasic and BackgroundActivity) 
 ### :bug: Bug Fixes
 * Do not register SIGHUP signal in windows